# 01_Backend/storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'API Documentation - ThuanNC'
  description: 'L5 Swagger OpenApi description'
  contact:
    email: congthuan701@gmail.com
  version: v1.0
servers:
  -
    url: 'http://thuan-be:2001/api'
    description: 'Laravel & PostgreSQL'
paths:
...
  /tour:
    get:
      tags:
        - Tour
      summary: 'Search tour'
      description: "\n     *      Code\n     *          200 - Success\n     *          400 - Bad request\n     *          401 - Not authentication\n     *          403 - Not access\n     *          422 - Input invalidate\n     *          423 - Have other error\n     *          500 - Server error\n     *      "
      operationId: 1988c4fdfca71dd358983e7c99415901
      parameters:
        -
          name: title
          in: query
          description: 'Tên tour du lịch'
          schema:
            type: string
        -
          name: departure_time
          in: query
          description: 'Ngày đi'
          schema:
            type: string
        -
          name: arrival_time
          in: query
          description: 'Ngày về'
          schema:
            type: string
        -
          name: adult_ticket_price
          in: query
          description: 'Giá tour'
          schema:
            type: string
        -
          name: offset
          in: query
          description: 'Số lượng hiển thị'
          schema:
            type: int
        -
          name: currentPage
          in: query
          description: 'Chọn page hiện tại'
          schema:
            type: int
      responses:
        200:
          description: 'Result of success'
          content:
            application/json:
              schema:
                properties:
                  Code:
                    type: integer
                    example: '200'
                  Data:
                    description: 'Result of success'
                type: object
      security:
        -
          apiAuth: []
    post:
      tags:
        - Tour
      summary: 'Save tour'
      description: "\n     *      Code\n     *          200 - Success\n     *          400 - Bad request\n     *          401 - Not authentication\n     *          403 - Not access\n     *          422 - Input invalidate\n     *          423 - Have other error\n     *          500 - Server error\n     *      "
      operationId: 1f313705a40fbbccb5c341539f0dc6ff
      requestBody:
        description: 'Start date'
        content:
          application/json:
            schema:
              properties:
                tourData:
                  type: object
                  example: '{...}'
                tourDateData:
                  type: arr
                  example: '[...]'
              type: object
      responses:
        200:
          description: 'Result of success'
          content:
            application/json:
              schema:
                properties:
                  Code:
                    type: integer
                    example: '200'
                  Data:
                    description: 'Result of success'
                type: object
      security:
        -
          apiAuth: []
...
components:
  securitySchemes:
    apiAuth:
      type: http
      description: 'Login with email and password to get the authentication token'
      name: 'Token based Based'
      in: header
      bearerFormat: JWT
      scheme: bearer
...